Hotfix random seed

+++ include/remollRunData.hh
@@ -32,7 +32,7 @@ class remollRunData : public TObject {
 	const char *GetGenName(){ return fGenName; }
 
 	void SetBeamE(double E){ fBeamE = E; }
-	void SetSeed(unsigned int seed){ fSeed = seed; }
+	void SetSeed(unsigned long int seed){ fSeed = seed; }
 
 	void AddMagData(filedata_t d){fMagData.push_back(d);}
 	void SetMacroFile(const char *fn){ fMacro = remollTextFile(fn); }
@@ -48,7 +48,7 @@ class remollRunData : public TObject {
 	TTimeStamp fRunTime;
 
 	long int  fNthrown;
-	unsigned int  fSeed;
+	long int  fSeed;
 	double fBeamE;
 	char fGenName[__RUNSTR_LEN];
